Add files via upload
The projects here include the attached files, which consists of a Proteus Software file, where a schematic is used to run a program using a code and is used to run the Stepper Motor. The .com file is a very important file which has to be added inside the very first IC of 8086.